refactor: move ibp/socket communication into runner/pkg/
