+++ runner/pkg/socket/BUILD.bazel
@@ -6,6 +6,6 @@ go_library(
         "json.go",
         "socket.go",
     ],
-    importpath = "github.com/aspect-build/aspect-gazelle/common/socket",
+    importpath = "github.com/aspect-build/aspect-gazelle/runner/pkg/socket",
     visibility = ["//visibility:public"],
 )
