+++ runner/pkg/watchman/watch.go
@@ -11,7 +11,7 @@ import (
 	"sync/atomic"
 
 	"github.com/aspect-build/aspect-gazelle/common/bazel"
-	"github.com/aspect-build/aspect-gazelle/common/socket"
+	"github.com/aspect-build/aspect-gazelle/runner/pkg/socket"
 )
 
 type ChangeSet struct {
